Feature ETP-3674: Split DummyHttpServletRequest to fix Sonar method limit
