+++ src/com/etendoerp/asyncprocess/startup/DummyHttpServletRequest.java
@@ -1,133 +1,49 @@
 package com.etendoerp.asyncprocess.startup;
 
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import java.security.Principal;
-import java.util.Enumeration;
+import jakarta.servlet.http.HttpUpgradeHandler;
+import jakarta.servlet.http.Part;
+import java.util.Collections;
 
 /**
- * Lightweight dummy implementation of {@link HttpServletRequest} intended for use in
- * asynchronous/background contexts where a full servlet request object is not available.
+ * Lightweight dummy implementation of {@link jakarta.servlet.http.HttpServletRequest} intended
+ * for use in asynchronous/background contexts where a full servlet request is not available.
  *
- * <p>This class provides safe, minimal default values for commonly used methods so
- * that code which expects an HttpServletRequest can execute without requiring a live
- * HTTP request (for example when constructing messages or simulating a request in
- * asynchronous processing).</p>
+ * <p>This class provides safe, minimal default values for commonly used methods so that code
+ * which expects an {@code HttpServletRequest} can execute without a live HTTP request — for
+ * example when constructing messages or simulating a request in asynchronous processing.</p>
  *
  * <p>Behavior and limitations:
  * <ul>
- *   <li>Most mutation and session-related operations are no-ops and return null or
- *       empty collections.</li>
- *   <li>Default values are intentionally simple and deterministic (e.g. getMethod()
- *       returns "POST", getRequestURI() returns "/async-process").</li>
- *   <li>This class is not a full-featured HttpServletRequest implementation and must
- *       not be used for real HTTP request processing.</li>
+ *   <li>Most mutation and session-related operations are no-ops and return null or empty
+ *       collections.</li>
+ *   <li>Default values are intentionally simple and deterministic (e.g. {@link #getMethod()}
+ *       returns "POST", {@link #getRequestURI()} returns "/async-process").</li>
+ *   <li>This class is not a full-featured implementation and must not be used for real HTTP
+ *       request processing.</li>
  * </ul>
  * </p>
+ *
+ * <p>Implementation is split across three classes to comply with the 35-method class limit:
+ * {@link AbstractDummyServletRequest} → {@link AbstractDummyHttpServletRequest} →
+ * {@code DummyHttpServletRequest}.</p>
  */
-class DummyHttpServletRequest implements HttpServletRequest {
-  @Override
-  public String getAuthType() {
-    return null;
-  }
-
-  @Override
-  public Cookie[] getCookies() {
-    return new Cookie[0];
-  }
-
-  @Override
-  public long getDateHeader(String name) {
-    return -1;
-  }
-
-  @Override
-  public String getHeader(String name) {
-    return null;
-  }
-
-  @Override
-  public Enumeration<String> getHeaders(String name) {
-    return java.util.Collections.emptyEnumeration();
-  }
-
-  @Override
-  public Enumeration<String> getHeaderNames() {
-    return java.util.Collections.emptyEnumeration();
-  }
-
-  @Override
-  public int getIntHeader(String name) {
-    return -1;
-  }
-
-  /**
-   * Returns the simulated HTTP method for this dummy request.
-   * Default: "POST".
-   */
-  @Override
-  public String getMethod() {
-    return "POST";
-  }
-
-  @Override
-  public String getPathInfo() {
-    return null;
-  }
-
-  @Override
-  public String getPathTranslated() {
-    return null;
-  }
-
-  /**
-   * Returns the servlet context path for the simulated request.
-   * Default: empty string to represent the application root context.
-   */
-  @Override
-  public String getContextPath() {
-    return "";
-  }
-
-  @Override
-  public String getQueryString() {
-    return null;
-  }
-
-  @Override
-  public String getRemoteUser() {
-    return null;
-  }
-
-  @Override
-  public boolean isUserInRole(String role) {
-    return false;
-  }
-
-  @Override
-  public Principal getUserPrincipal() {
-    return null;
-  }
+class DummyHttpServletRequest extends AbstractDummyHttpServletRequest {
 
   @Override
   public String getRequestedSessionId() {
     return null;
   }
 
-  /**
-   * Returns a fixed request URI used by asynchronous processing components.
-   * Default: "/async-process".
-   */
+  /** Returns a fixed request URI used by asynchronous processing. Default: "/async-process". */
   @Override
   public String getRequestURI() {
     return "/async-process";
   }
 
   /**
-   * Returns a simple request URL representation. This is a convenience value and
-   * should not be treated as a full URL builder.
+   * Returns a simple request URL representation.
    * Default: "http://localhost:8080/etendo".
    */
   @Override
@@ -135,10 +51,7 @@ class DummyHttpServletRequest implements HttpServletRequest {
     return new StringBuffer("http://localhost:8080/etendo");
   }
 
-  /**
-   * Returns the servlet path for the simulated request.
-   * Default: "/async-process".
-   */
+  /** Returns the servlet path for the simulated request. Default: "/async-process". */
   @Override
   public String getServletPath() {
     return "/async-process";
@@ -174,6 +87,7 @@ class DummyHttpServletRequest implements HttpServletRequest {
     return false;
   }
 
+  @Override
   public boolean isRequestedSessionIdFromUrl() {
     return false;
   }
@@ -192,235 +106,17 @@ class DummyHttpServletRequest implements HttpServletRequest {
   }
 
   @Override
-  public java.util.Collection<jakarta.servlet.http.Part> getParts() {
-    return java.util.Collections.emptyList();
+  public java.util.Collection<Part> getParts() {
+    return Collections.emptyList();
   }
 
   @Override
-  public jakarta.servlet.http.Part getPart(String name) {
+  public Part getPart(String name) {
     return null;
   }
 
   @Override
-  public <T extends jakarta.servlet.http.HttpUpgradeHandler> T upgrade(Class<T> handlerClass) {
+  public <T extends HttpUpgradeHandler> T upgrade(Class<T> handlerClass) {
     return null;
-  }
-
-  @Override
-  public Object getAttribute(String name) {
-    return null;
-  }
-
-  @Override
-  public Enumeration<String> getAttributeNames() {
-    return java.util.Collections.emptyEnumeration();
-  }
-
-  /**
-   * Returns the content type that should be assumed for the dummy request.
-   * Default: "application/json".
-   */
-  @Override
-  public String getContentType() {
-    return "application/json";
-  }
-
-  /**
-   * Returns the character encoding used by this dummy request.
-   * Default: "UTF-8".
-   */
-  @Override
-  public String getCharacterEncoding() {
-    return "UTF-8";
-  }
-
-  @Override
-  public void setCharacterEncoding(String env) {
-  }
-
-  @Override
-  public int getContentLength() {
-    return -1;
-  }
-
-  @Override
-  public long getContentLengthLong() {
-    return -1L;
-  }
-
-  @Override
-  public jakarta.servlet.ServletInputStream getInputStream() {
-    return null;
-  }
-
-  @Override
-  public String getParameter(String name) {
-    return null;
-  }
-
-  @Override
-  public Enumeration<String> getParameterNames() {
-    return java.util.Collections.emptyEnumeration();
-  }
-
-  @Override
-  public String[] getParameterValues(String name) {
-    return null;
-  }
-
-  @Override
-  public java.util.Map<String, String[]> getParameterMap() {
-    return java.util.Collections.emptyMap();
-  }
-
-  @Override
-  public String getProtocol() {
-    return "HTTP/1.1";
-  }
-
-  @Override
-  public String getScheme() {
-    return "http";
-  }
-
-  /**
-   * Returns the host name used in the dummy request environment.
-   * Default: "localhost".
-   */
-  @Override
-  public String getServerName() {
-    return "localhost";
-  }
-
-  /**
-   * Returns the server port used in the dummy request environment.
-   * Default: 8080.
-   */
-  @Override
-  public int getServerPort() {
-    return 8080;
-  }
-
-  @Override
-  public java.io.BufferedReader getReader() {
-    return null;
-  }
-
-  /**
-   * Returns the remote IP address associated with this dummy request.
-   * Default: "127.0.0.1".
-   */
-  @Override
-  public String getRemoteAddr() {
-    return "127.0.0.1";
-  }
-
-  @Override
-  public String getRemoteHost() {
-    return "localhost";
-  }
-
-  @Override
-  public void setAttribute(String name, Object o) {
-  }
-
-  @Override
-  public void removeAttribute(String name) {
-  }
-
-  @Override
-  public java.util.Locale getLocale() {
-    return java.util.Locale.getDefault();
-  }
-
-  @Override
-  public Enumeration<java.util.Locale> getLocales() {
-    return java.util.Collections.enumeration(
-        java.util.Arrays.asList(java.util.Locale.getDefault()));
-  }
-
-  @Override
-  public boolean isSecure() {
-    return false;
-  }
-
-  @Override
-  public jakarta.servlet.RequestDispatcher getRequestDispatcher(String path) {
-    return null;
-  }
-
-  public String getRealPath(String path) {
-    return null;
-  }
-
-  @Override
-  public String getRequestId() {
-    return "";
-  }
-
-  @Override
-  public String getProtocolRequestId() {
-    return "";
-  }
-
-  @Override
-  public jakarta.servlet.ServletConnection getServletConnection() {
-    return null;
-  }
-
-  @Override
-  public int getRemotePort() {
-    return 0;
-  }
-
-  @Override
-  public String getLocalName() {
-    return "localhost";
-  }
-
-  @Override
-  public String getLocalAddr() {
-    return "127.0.0.1";
-  }
-
-  @Override
-  public int getLocalPort() {
-    return 8080;
-  }
-
-  @Override
-  public jakarta.servlet.ServletContext getServletContext() {
-    return null;
-  }
-
-  @Override
-  public jakarta.servlet.AsyncContext startAsync() {
-    return null;
-  }
-
-  @Override
-  public jakarta.servlet.AsyncContext startAsync(jakarta.servlet.ServletRequest servletRequest,
-      jakarta.servlet.ServletResponse servletResponse) {
-    return null;
-  }
-
-  @Override
-  public boolean isAsyncStarted() {
-    return false;
-  }
-
-  @Override
-  public boolean isAsyncSupported() {
-    return false;
-  }
-
-  @Override
-  public jakarta.servlet.AsyncContext getAsyncContext() {
-    return null;
-  }
-
-  @Override
-  public jakarta.servlet.DispatcherType getDispatcherType() {
-    return jakarta.servlet.DispatcherType.REQUEST;
   }
 }
